모델_배포/requirements.txt
tensorflow
tensorflow-serving-api
request